рефактор виджетов нотификации и модалок результатов отправки на сервер
